Upload file using BufferedReader, better for large files

+++ Nightfall.Net/Nightfall.Net/Models/Requests/InitiateUploadRequest.cs
@@ -5,13 +5,13 @@ namespace Nightfall.Net.Models.Requests
     public class InitiateUploadRequest
     {
         [JsonPropertyName("fileSizeBytes")]
-        public int FileSizeBytes { get; set; }
+        public long FileSizeBytes { get; set; }
 
         public InitiateUploadRequest()
         {
             
         }
-        public InitiateUploadRequest(int fileSizeBytes)
+        public InitiateUploadRequest(long fileSizeBytes)
         {
             FileSizeBytes = fileSizeBytes;
         }
